+++ changelog.md
@@ -1,2 +1,4 @@
 # Changes
-- Updating CC: Advanced Math to 0.3.5
+- New block details under `sable` key!
+- Updating CC: Advanced Math to 0.3.6!
+- Updating documentation and README!